MoviesPresenter moving repository package to domain
