+++ app/src/main/java/com/tomerpacific/moviepresenter/domain/model/MovieModel.kt
@@ -1,4 +1,4 @@
-package com.tomerpacific.moviepresenter.domain
+package com.tomerpacific.moviepresenter.domain.model
 
 import android.graphics.Bitmap
 import kotlinx.serialization.SerialName
