+++ app/src/main/java/com/tomerpacific/moviepresenter/model/TMDBResponse.kt
@@ -1,6 +1,6 @@
 package com.tomerpacific.moviepresenter.model
 
-import com.tomerpacific.moviepresenter.domain.MovieModel
+import com.tomerpacific.moviepresenter.domain.model.MovieModel
 import kotlinx.serialization.Serializable
 
 @Serializable
